city ekleme silme ve güncelleme işlemleri için contoller ve view model oluşturuldu gerekli işlemler yapıldı

+++ ActivityOrganizerManager/OrganizationActivityManager/ViewModels/CategoryViewModel.cs
@@ -4,7 +4,7 @@ namespace AcitivityOrganizerManager.ViewModels
 {
     public class CategoryViewModel
     {
-        public int CategoryId { get; set; }
+        public byte CategoryId { get; set; }
         public string CategoryName { get; set; }
